Move next and prev buttons to folders
