+++ src/containers/play-bar/desktop/PlayBar.js
@@ -31,8 +31,8 @@ import { getTheme } from 'store/application/ui/theme/selectors'
 import PlayingTrackInfo from './components/PlayingTrackInfo'
 import VolumeBar from 'components/play-bar/VolumeBar'
 import PlayButton from 'components/play-bar/PlayButton'
-import NextButton from 'components/play-bar/NextButton'
-import PreviousButton from 'components/play-bar/PreviousButton'
+import NextButton from 'components/play-bar/next-button/NextButton'
+import PreviousButton from 'components/play-bar/previous-button/PreviousButton'
 import ShuffleButtonProvider from 'components/play-bar/shuffle-button/ShuffleButtonProvider'
 import RepeatButtonProvider from 'components/play-bar/repeat-button/RepeatButtonProvider'
 import RepostButton from 'components/general/RepostButton'
