+++ src/stories/play-bar/index.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 
 import PlayButton from 'components/play-bar/PlayButton'
-import NextButton from 'components/play-bar/NextButton'
-import PreviousButton from 'components/play-bar/PreviousButton'
+import NextButton from 'components/play-bar/next-button/NextButton'
+import PreviousButton from 'components/play-bar/previous-button/PreviousButton'
 import RepeatButton from 'components/play-bar/repeat-button/RepeatButton'
 import ShuffleButton from 'components/play-bar/shuffle-button/ShuffleButton'
 import VolumeBar from 'components/play-bar/VolumeBar'
